address more PR comments
